FaceRecognizer/requirements.txt
fastapi
uvicorn
face_recognition
numpy
pillow
python-multipart
boto3
python-dotenv
